Implemented basic primitive-based box-model rendering for TextPrimitive and ImagePrimitive. Remove CairoPrimitiveRenderer: not needed any more since the rendering theory requires GPU-based rendering.

+++ src/ImGui/OSAbstraction/Text/ITextContext.cs
@@ -11,7 +11,7 @@ namespace ImGui.OSAbstraction.Text
         /// <summary>
         /// size of the font in the text
         /// </summary>
-        float FontSize { get; }
+        double FontSize { get; }
 
         /// <summary>
         /// alignment of the text in the rectangle
